backend/requirements.txt
asgiref==3.7.2
asttokens==2.4.1
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==2.0.12
colorama==0.4.6
decorator==5.1.1
Django==3.1.12
django-cors-headers==3.11.0
djangorestframework==3.14.0
exceptiongroup==1.1.3
executing==2.0.1
idna==3.4
ipython==8.17.2
jedi==0.19.1
matplotlib-inline==0.1.6
numpy==1.26.1
packaging==23.2
pandas==2.1.2
parso==0.8.3
plotly==5.18.0
prompt-toolkit==3.0.39
pure-eval==0.2.2
Pygments==2.16.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.27.1
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
tenacity==8.2.3
tqdm==4.64.1
traitlets==5.13.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
wcwidth==0.2.9
python-decouple==3.8
djongo==1.3.6
pymongo==3.12.1
